+++ nsd/queue.c
@@ -1011,7 +1011,7 @@ NsConnThread(void *arg)
     NsServer      *servPtr = poolPtr->servPtr;
     Conn          *connPtr = NULL;
     Ns_Time        wait, *timePtr = &wait;
-    unsigned int   id;
+    uintptr_t      id;
     bool           shutdown;
     int            status = NS_OK, cpt, ncons, current, fromQueue;
     long           timeout;
@@ -1042,9 +1042,10 @@ NsConnThread(void *arg)
     assert(poolPtr->pool != NULL);
     { 
 	const char *p = *poolPtr->pool != '\0' ? poolPtr->pool : NULL;
-	Ns_ThreadSetName("-conn:%s%s%s:%d", servPtr->server, 
+	Ns_ThreadSetName("-conn:%s%s%s:%" PRIuPTR, servPtr->server, 
 			 (p != NULL) ? ":" : "", 
-			 (p != NULL) ? p : "", id);
+			 (p != NULL) ? p : "",
+                         id);
     }
 
     /*
